requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
qdrant-client
sentence-transformers
transformers
torch
openai
groq
sqlalchemy
alembic
python-multipart
aiofiles
httpx
numpy
scikit-learn
faiss-cpu
rank-bm25
python-dotenv
loguru
prometheus-client
pytest
aiosqlite
celery[redis]
redis
kombu
flower
# Circuit breaker and resilience

tenacity
async-timeout
# Load testing
locust
aiohttp
asyncio-throttle
# Monitoring and metrics
psutil
memory-profiler 